+++ src/collective/upgrade/run.py
@@ -1,6 +1,6 @@
 import sys
 import logging
-import optparse
+import argparse
 import pdb
 
 import transaction
@@ -9,20 +9,20 @@ import zodbupdate.main
 from collective.upgrade import utils
 
 
-parser = optparse.OptionParser()
-parser.add_option(
+parser = argparse.ArgumentParser()
+parser.add_argument(
     "-l", "--log-file", metavar="FILE", default='upgrade.log',
     help="Log upgrade messages, filtered for duplicates, to FILE")
-parser.add_option(
+parser.add_argument(
     "-p", "--portal-path", metavar="PATH", action="append",
     help="Run upgrades for the portals at the given paths only.  "
     "May be given multiple times to specify multiple portals.  "
     "If not given, all CMF portals in the Zope app will be upgraded.")
-parser.add_option(
+parser.add_argument(
     "-z", "--zope-conf", metavar="FILE",
     help='The "zope.conf" FILE to use when starting the Zope2 app. '
     'Can be omitted when used as a zopectl "run" script.')
-parser.add_option(
+parser.add_argument(
     "-d", "--disable-link-integrity", action="store_true",
     help='When upgrading a portal using plone.app.linkintegrity, '
     'disable it during the upgrade.')
@@ -32,21 +32,19 @@ def main(app=None, args=None):
     full_args = args
     if args is not None:
         full_args = args + sys.argv[1:]
-    options, args = parser.parse_args(full_args)
-    if args:
-        parser.error('Unrecognized args given: %r' % args)
+    args = parser.parse_args(full_args)
 
     if app is None:
         import Zope2
         from App import config
         if config._config is None:
-            if not options.zope_conf:
+            if not args.zope_conf:
                 parser.error(
                     'Must give the "--zope-conf" option when not used as a '
                     'zopectl "run" script.')
-            Zope2.configure(options.zope_conf)
+            Zope2.configure(args.zope_conf)
         app = Zope2.app()
-    elif options.zope_conf:
+    elif args.zope_conf:
         parser.error(
             'Do not give the "--zope-conf" option when used as a '
             'zopectl "run" script.')
@@ -58,13 +56,13 @@ def main(app=None, args=None):
     stderr_handler.setLevel(logging.INFO)
     stderr_handler.addFilter(zodbupdate.main.duplicate_filter)
 
-    log_file = logging.FileHandler(options.log_file)
+    log_file = logging.FileHandler(args.log_file)
     log_file.addFilter(zodbupdate.main.duplicate_filter)
     log_file.setFormatter(utils.formatter)
     root.addHandler(log_file)
 
-    kw = dict(paths=options.portal_path)
-    if options.disable_link_integrity:
+    kw = dict(paths=args.portal_paths)
+    if args.disable_link_integrity:
         kw['enable_link_integrity_checks'] = False
 
     from AccessControl import SpecialUsers